+++ react_on_rails/spec/dummy/config/routes.rb
@@ -30,7 +30,6 @@ Rails.application.routes.draw do
   get "client_side_manual_render" => "pages#client_side_manual_render"
   get "render_js" => "pages#render_js"
   get "react_router(/*all)" => "react_router#index", as: :react_router
-  get "tanstack_router(/*all)" => "tanstack_router#index", as: :tanstack_router
   get "pure_component" => "pages#pure_component"
   get "css_modules_images_fonts_example" => "pages#css_modules_images_fonts_example"
   get "turbolinks_cache_disabled" => "pages#turbolinks_cache_disabled"